+++ R/ef_whe.R
@@ -16,27 +16,27 @@
 #' @return An emission factor by annual mileage.
 #' @keywords emission factors high emitters
 #' @export
-#' @examples \dontrun{
+#' @examples {
 #' # Do not run
 #' # Let's say high emitter is 5 times the normal ef.
 #' co_efhe <- ef_cetesb(p = "COd", "PC_G") * 5
 #' # Let's say that the perfil of high emitters increases linearly
 #' # till 30 years and after that percentage is constant
-#' perc <- c(seq(0.01, 0.3, 0.01), rep(0.3, 20))
+#' perc <- c(seq(0.01, 0.3, 0.01), rep(0.3, 10))
 #' # Now, lets use our ef with normal deterioration
 #' co_ef_normal <- ef_cetesb(p = "COd", "PC_G")
-#' efd <- ef_whe(efhe = co_efhe, phe = perc, ef = co_ef_normal)
+#' efd <- ef_whe(efhe = co_efhe,
+#'               phe = perc,
+#'               ef = co_ef_normal)
 #' # now, we can plot the three ef
-#' plot(co_efhe)
-#' lines(co_ef_normal, pch = 16, col = "red" )
-#' lines(efd, pch = 16, col = "blue")
+#' colplot(data.frame(co_efhe, co_ef_normal, efd))
 #' }
 ef_whe <- function(efhe,
-                  phe,
-                  ef){
+                   phe,
+                   ef){
   if(missing(efhe)) stop("No high emitter emission factor 'he'")
   if(missing(phe)) stop("No percentage of high emitter vehicles 'phe'")
   if(missing(ef)) stop("No emission factors with normal deterioration 'ef'")
   eff <- as.numeric(efhe) * as.numeric(phe) + as.numeric(ef) * (1 - phe)
-return(EmissionFactors(eff))
+  return(EmissionFactors(eff))
 }