+++ tests/testthat/test-emis_evap.R
@@ -4,14 +4,18 @@ data(net)
 a <- Vehicles(1:10)
 lkm <- units::as_units(1:10, "km")
 ef <- EmissionFactors(1:10)
-netsf <- sf::st_as_sf(net)[1:10, "ldv"]
+
+netsf <- net[1:10, "ldv"]
 
 test_that("emis_evap works", {
   expect_equal(emis_evap(veh = a, x = lkm, hotfi = ef,
                          verbose = TRUE)[1],
                Emissions(1))
-  expect_message(emis_evap(veh = a, x = lkm, hotfi = ef, verbose = TRUE),
-                 "E.?")
+  # expect_message(emis_evap(veh = a,
+  #                          x = lkm,
+  #                          hotfi = ef,
+  #                          verbose = TRUE),
+  #                ".?")
   expect_equal(emis_evap(veh = netsf$ldv,
                          x = lkm,
                          hotfi = ef,
@@ -38,11 +42,11 @@ test_that("emis_evap works", {
                            x = lkm,
                            ed = data.frame(Emissions(1)))$emissions[1],
                  Emissions(4350))
-  expect_message(emis_evap(veh = netsf,
-                         x = lkm,
-                         ed = data.frame(Emissions(1)),
-                         verbose = TRUE),
-               "E.?")
+  # expect_message(emis_evap(veh = netsf,
+  #                        x = lkm,
+  #                        ed = data.frame(Emissions(1)),
+  #                        verbose = TRUE),
+  #              "E.?")
   expect_output(emis_evap(veh = netsf,
                            x = lkm,
                            ed = data.frame(Emissions(1)),
@@ -68,11 +72,11 @@ test_that("emis_evap works", {
                          x = lkm,
                          ed = dfef)$emissions[1],
                Emissions(1))
-  expect_message(emis_evap(veh = dfa,
-                         x = lkm,
-                         ed = dfef,
-                         verbose = TRUE)$emissions[1],
-                "E.?")
+  # expect_message(emis_evap(veh = dfa,
+  #                        x = lkm,
+  #                        ed = dfef,
+  #                        verbose = TRUE)$emissions[1],
+  #               "E.?")
   expect_output(print(emis_evap(veh = dfa,
                            x = lkm,
                            ed = dfef,
@@ -84,12 +88,12 @@ test_that("emis_evap works", {
                          pro_month = 1:12,
                          verbose = TRUE)$emissions[1]),
                Emissions(0))
-  expect_message(emis_evap(veh = dfa,
-                               x = lkm,
-                               ed = dfefm,
-                               pro_month = 1:12,
-                               verbose = TRUE)$emissions[1],
-               ".?")
+  # expect_message(emis_evap(veh = dfa,
+  #                              x = lkm,
+  #                              ed = dfefm,
+  #                              pro_month = 1:12,
+  #                              verbose = TRUE)$emissions[1],
+  #              ".?")
 
   expect_error(emis_evap(veh = dfa,
                            x = lkm,
@@ -115,14 +119,14 @@ test_that("emis_evap works", {
                          x = lkm,
                          carb = 0.5)[1],
                ".?")
-  expect_message(emis_evap(veh = a,
-                         x = lkm,
-                         p = 0.1,
-                         hotc = 10,
-                         warmc = 10,
-                         hotfi = 0.1,
-                         carb = 0.5, verbose = TRUE)[1],
-               ".?")
+  # expect_message(emis_evap(veh = a,
+  #                        x = lkm,
+  #                        p = 0.1,
+  #                        hotc = 10,
+  #                        warmc = 10,
+  #                        hotfi = 0.1,
+  #                        carb = 0.5, verbose = TRUE)[1],
+  #              ".?")
   expect_equal(emis_evap(veh = a,
                            x = lkm,
                            p = 0.1,
@@ -143,15 +147,15 @@ test_that("emis_evap works", {
                          verbose = TRUE)$emissions[1],
                Emissions(1))
 
-  expect_message(emis_evap(veh = dfa,
-                         x = lkm,
-                         p = 0.1,
-                         hotc = dfa,
-                         warmc = dfa,
-                         hotfi = dfa,
-                         carb = 2,
-                         verbose = TRUE),
-               ".?")
+  # expect_message(emis_evap(veh = dfa,
+  #                        x = lkm,
+  #                        p = 0.1,
+  #                        hotc = dfa,
+  #                        warmc = dfa,
+  #                        hotfi = dfa,
+  #                        carb = 2,
+  #                        verbose = TRUE),
+  #              ".?")
 
   expect_output(emis_evap(veh = dfa,
                            x = lkm,
@@ -180,14 +184,14 @@ test_that("emis_evap works", {
                          x = lkm,
                          carb = 0.5)[1],
                ".?")
-  expect_message(emis_evap(veh = dfa,
-                           x = lkm,
-                           p = 0.1,
-                           hotc = data.frame(matrix(10, ncol = ncol(dfa))),
-                           warmc = data.frame(matrix(10, ncol = ncol(dfa))),
-                           hotfi = data.frame(matrix(10, ncol = ncol(dfa))),
-                           carb = 0.5, verbose = TRUE)[1],
-                 ".?")
+  # expect_message(emis_evap(veh = dfa,
+  #                          x = lkm,
+  #                          p = 0.1,
+  #                          hotc = data.frame(matrix(10, ncol = ncol(dfa))),
+  #                          warmc = data.frame(matrix(10, ncol = ncol(dfa))),
+  #                          hotfi = data.frame(matrix(10, ncol = ncol(dfa))),
+  #                          carb = 0.5, verbose = TRUE)[1],
+  #                ".?")
   expect_equal(emis_evap(veh = dfa,
                            x = lkm,
                            p = 0.1,