+++ NEWS.md
@@ -1,6 +1,10 @@
 NEWS
 ===========
 
+### vein 1.0.2
+
+- Change all units according units package version 0.8.2.
+- Deprecated some functions.
 
 ### vein 1.0.1